Classess
Write class declarations for the following class hierarchy. A base class
(parent
class) named Student that contains the identification string “Student”
along with a
virtual printOutput() function that displays the string. The derived
classes (child
class) Fulltime and PartTime each contain a string identifying the type
of Student.
The function printOutput() overrides printOutput() in the base class. It
outputs
the string “This person is a” and calls the base class function
printOutput().
Test the hierarchy by declaring an array, studentTypeList 3 pointers to
the base.
The program creates an appropriate Student object based on input. The
user
type ‘f’ or ‘p’ to indicate the type of Student. The array is
initialized by
dynamically allocating objects based on user input. The program then
calls the
printOutput() function for each object.